requirements.txt
SQLAlchemy==2.0.23
uvicorn==0.24.0.post1
fastapi==0.105.0
pydantic==2.5.2
py-bcrypt==0.4
cryptography==41.0.7
python-jose==3.3.0
